refactor: Edit files and directories names to follow naming convention (dash-case)
