+++ src/pages/event-dashboard/action-list/ActionList.jsx
@@ -1,4 +1,4 @@
-import './actionList.css';
+import './action-list.css'
 
 function ActionList(props) {
     return (
