+++ src/pages/event-dashboard/card/card.css
